moondream/requirements.txt
moondream
Pillow 